backend/requirements.txt
fastapi==0.101.1
uvicorn==0.22.0
assemblyai==0.41.1
python-dotenv==1.0.0
python-multipart==0.0.20

